+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Implemented enhancements
 
 - No ticket - Moved over GDS PaaS S3 bucket.
+- TT-1574 - Support more non-companies house companies
 - TT-1590 - Port ISD search logic to FAS
 - No ticket - Remove obsolete case study search endpoint
 - No ticket -  Mask product data make email unique